feat: Update frontend dependencies
The code changes in the frontend/package.json file update the dependencies for the project. The @paypal/react-paypal-js library is added to enable PayPal integration in the Checkout component. The css-loader package is also added to handle CSS loading. This change improves the functionality and user experience of the application.