```text refactor: Shuffle advertisements in Home component
This code change shuffles the advertisements before setting them to state in the Home component. By randomly reordering the advertisements, we provide a more dynamic and engaging experience for users visiting the home page. This enhancement improves the visual presentation of the featured properties and adds an element of surprise to the carousel.